Update global and patient components
And move them to shared components
