+++ src/js/views/globals/search/patient-search.scss
@@ -57,7 +57,7 @@
 .patient-search__picklist-item {
   color: $black-20;
   font-size: 16px;
-  padding: 0 40px;
+  padding: 4px 56px;
 }
 
 .picklist-item__birthdate {
