--- src/js/views/patients/actions/actions.scss
@@ -1,31 +1,0 @@
-.actions__owner {
-  display: inline-block;
-  max-width: 80px;
-}
-
-.actions__role {
-  color: $black-60;
-  font-size: 12px;
-  margin-left: 4px;
-}
-
-.is-highlighted .actions__role {
-  color: $white;
-}
-
-.actions__due .is-overdue {
-  color: color(red);
-}
-
-.actions__due {
-  border-radius: 4px 0 0 4px;
-}
-
-.actions__time {
-  border-left: 1px solid transparent;
-  border-radius: 0 4px 4px 0;
-
-  &:hover {
-    border-left: 1px solid $black-90;
-  }
-}
