+++ src/js/views/patients/shared/actions_views.js
@@ -1,0 +1,32 @@
+import Radio from 'backbone.radio';
+import hbs from 'handlebars-inline-precompile';
+import { View } from 'marionette';
+
+import 'sass/modules/buttons.scss';
+
+import StateComponent from './components/state_component';
+import OwnerComponent from './components/owner_component';
+import DueComponent from './components/due_component';
+import TimeComponent from './components/time_component';
+import DurationComponent from './components/duration_component';
+
+const AttachmentButton = View.extend({
+  className: 'button-secondary--compact is-icon-only',
+  tagName: 'button',
+  template: hbs`{{far "link"}}`,
+  triggers: {
+    'click': 'click',
+  },
+  onClick() {
+    Radio.trigger('event-router', 'form:patientAction', this.model.id, this.model.getForm().id);
+  },
+});
+
+export {
+  StateComponent,
+  OwnerComponent,
+  DueComponent,
+  TimeComponent,
+  DurationComponent,
+  AttachmentButton,
+};
