+++ src/js/views/patients/shared/components/due-component.scss
@@ -1,0 +1,8 @@
+.due-component .is-overdue {
+  color: color(red);
+}
+
+.due-component {
+  border-radius: 4px 0 0 4px;
+}
+
