+++ src/js/views/patients/shared/flows_views.js
@@ -1,0 +1,53 @@
+import Radio from 'backbone.radio';
+
+import 'sass/modules/buttons.scss';
+
+import intl from 'js/i18n';
+
+import StateComponent from './components/state_component';
+import OwnerComponent from './components/owner_component';
+
+const i18n = intl.patients.shared.flowsViews;
+
+const FlowStateComponent = StateComponent.extend({
+  onPicklistSelect({ model }) {
+    // Selected done
+    if (model.get('status') === 'done' && this.getOption('flow')) {
+      this.shouldSelectDone(model);
+      return;
+    }
+
+    this.setSelectedStatus(model);
+  },
+  shouldSelectDone(model) {
+    const flow = this.getOption('flow');
+
+    if (flow.isAllDone()) {
+      this.setSelectedStatus(model);
+      return;
+    }
+
+    // We must hide the droplist before showing the modal
+    this.popRegion.empty();
+
+    const modal = Radio.request('modal', 'show:small', {
+      bodyText: i18n.doneModal.bodyText,
+      headingText: i18n.doneModal.headingText,
+      submitText: i18n.doneModal.submitText,
+      buttonClass: 'button--green',
+      onSubmit: () => {
+        this.setSelectedStatus(model);
+        modal.destroy();
+      },
+    });
+  },
+  setSelectedStatus(model) {
+    this.setState('selected', model);
+    this.popRegion.empty();
+  },
+});
+
+export {
+  FlowStateComponent,
+  OwnerComponent,
+};
